FileReader restructure
